app-controller/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
pyyaml>=6.0.1
pynvml>=11.5.0
redis>=5.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
requests>=2.31.0